Sprint 1: fix prod blockers (migration, money, auth, gate)
Fixes the highest-severity issues from the 2026-06-09 audit:

- Migration: remove duplicate CREATE TABLE permissions/role_permissions from
  9a847204f8d9 (already created by parent 692b91ac18be). Fresh 'alembic upgrade
  head' now runs clean end-to-end (verified base->head on a throwaway schema).
- Money: PaymentRequest.amount and DailyAccountBalance.amount Float -> Numeric(18,2);
  schemas use Decimal with field_serializer keeping JSON numeric; add ge=0
  validation; new migration 585af2e317c7; round liquidity/totals on the frontend
  before the surplus/deficit compare.
- Security: stop leaking hashed_password from create_user/update_user (return
  _serialize_user_safe); reject weak/placeholder SECRET_KEY at startup;
  whitelist+validate storage_path against path traversal.
- Gate: fail closed when the payment date has no calendar coverage or the org has
  no payment group (routes to PENDING_GATE), behind reversible
  GATE_REQUIRE_CALENDAR_COVERAGE flag (default true).

+++ app/models/balance.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime, timedelta, timezone
 
-from sqlalchemy import Column, Date, DateTime, Float, ForeignKey, String, Boolean, UniqueConstraint, Index
+from sqlalchemy import Column, Date, DateTime, Numeric, ForeignKey, String, Boolean, UniqueConstraint, Index
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
 
@@ -37,7 +37,7 @@ class DailyAccountBalance(Base):
     balance_date = Column(Date, nullable=False)
     organization_id = Column(UUID(as_uuid=True), ForeignKey("organizations.id"), nullable=False)
     bank_account_id = Column(UUID(as_uuid=True), ForeignKey("bank_accounts.id"), nullable=False)
-    amount = Column(Float, nullable=False)
+    amount = Column(Numeric(18, 2), nullable=False)
     created_by_id = Column(UUID(as_uuid=True), ForeignKey("users.id"), nullable=False)
     updated_by_id = Column(UUID(as_uuid=True), ForeignKey("users.id"), nullable=False)
     created_at = Column(DateTime, nullable=False, default=get_gmt3_time)
